requirements.txt
Flask
Werkzeug
pytest
PyJWT
Flask-Limiter
python-dotenv